+++ stag_lib/cluster.cpp
@@ -399,3 +399,23 @@ double stag::conductance(stag::LocalGraph* graph, std::vector<stag_int>& cluster
   if (volume == 0) return 0;
   else return cut / volume;
 }
+
+std::vector<stag_int> stag::symmetric_difference(std::vector<stag_int> &S,
+                                                 std::vector<stag_int> &T) {
+  // For the later steps, we assume that the vectors are sorted
+  std::sort(S.begin(), S.end());
+  std::sort(T.begin(), T.end());
+
+  // Remove duplicates from the vectors
+  auto last_idx = std::unique(S.begin(), S.end());
+  S.erase(last_idx, S.end());
+  last_idx = std::unique(T.begin(), T.end());
+  T.erase(last_idx, T.end());
+
+  // Compute and return the symmetric difference
+  std::vector<stag_int> symmetric_difference;
+  std::set_symmetric_difference(S.begin(), S.end(),
+                                T.begin(), T.end(),
+                                std::back_inserter(symmetric_difference));
+  return symmetric_difference;
+}